requirements.txt
aiosmtplib==2.0.2
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
async-generator==1.10
attrs==23.2.0
bcrypt==4.1.3
blinker==1.8.2
cffi==1.16.0
cfgv==3.4.0
click==8.1.7
colorlog==6.8.2
cryptography==42.0.5
distlib==0.3.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.1
fastapi-filter==2.0.0
fastapi-mail==1.4.1
fastapi-pagination==0.12.24
filelock==3.14.0
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httptools==0.6.1
identify==2.5.36
idna==3.7
Jinja2==3.1.4
Mako==1.3.3
MarkupSafe==2.1.5
mongoengine==0.27.0
nodeenv==1.8.0
outcome==1.3.0.post0
packaging==24.0
passlib==1.7.4
platformdirs==4.2.1
pre-commit==3.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.0
pydantic-settings==2.1.0
pydantic_core==2.18.1
PyJWT==2.8.0
pymongo==4.7.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.29
starlette==0.37.2
trio==0.21.0
typing_extensions==4.11.0
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.26.1
watchfiles==0.21.0
websockets==12.0
